requirements.txt
Django
djangorestframework
django-rest-auth[with_social]
django-extensions
djangorestframework-jwt
